pyproject.toml
[tool.poetry]
name = "azauto"
version = "0.1.0"
description = "A multipurpose webscraper"
authors = ["Joao Tedeschi <joaorafaelbt@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selectolax = "^0.3.14"
pandas="^1.3.3"
requests="^2.26.0"
numpy="^1.21.2"
python-dotenv="^1.0.0"
lxml = "^4.9.2"
html5lib = "^1.1"
bs4 = "^0.0.1"
rich = "^13.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
